Update csharp.md
Change text to 

Unity introduces the kids to the basics of object orientated programming in a fun manner. This course is suitable for students 10 years of age or older and who understand the basics of variables & objects, and who are at a Intermediate level of Scratch.

Unity uses a language called C# (pronounced C-sharp) . This is a scripting language so there is some typing involved. Students will need to download the software from https://unity.com/download and have it installed before coming to class.